requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
aiohttp>=3.9.0
pandas-ta>=0.3.14b0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pytest>=7.4.0
websockets>=10.0
scikit-learn>=1.0.0
lightgbm>=4.1.0